chore(patterns): split patterns breakdown
