+++ src/Components/ServiceScene/Breakdowns/PatternsBreakdownScene.tsx
@@ -1,33 +1,25 @@
 import { css } from '@emotion/css';
 import React from 'react';
 
-import { ConfigOverrideRule, DataFrame, FieldColor, FieldType, GrafanaTheme2, LoadingState } from '@grafana/data';
+import { DataFrame, FieldType, GrafanaTheme2 } from '@grafana/data';
 import {
   CustomVariable,
-  PanelBuilders,
   SceneComponentProps,
-  SceneCSSGridLayout,
-  SceneDataNode,
   SceneFlexItem,
   sceneGraph,
   SceneObject,
   SceneObjectBase,
   SceneObjectState,
   SceneVariableSet,
-  VizPanel,
 } from '@grafana/scenes';
-import { LegendDisplayMode, PanelContext, SeriesVisibilityChangeMode, Text, TextLink, useStyles2 } from '@grafana/ui';
+import { Text, TextLink, useStyles2 } from '@grafana/ui';
 import { LayoutSwitcher } from 'Components/ServiceScene/Breakdowns/LayoutSwitcher';
 import { StatusWrapper } from 'Components/ServiceScene/Breakdowns/StatusWrapper';
 import { GrotError } from 'Components/GrotError';
 import { VAR_LABEL_GROUP_BY } from 'services/variables';
 import { LokiPattern, ServiceScene } from '../ServiceScene';
-import { onPatternClick } from './FilterByPatternsButton';
 import { IndexScene } from '../../IndexScene/IndexScene';
-import { PatternsViewTableScene } from './PatternsViewTableScene';
-import { config } from '@grafana/runtime';
-
-const palette = config.theme2.visualization.palette;
+import { PatternsFrameScene } from './PatternsFrameScene';
 
 export interface PatternsBreakdownSceneState extends SceneObjectState {
   body?: SceneObject;
@@ -35,7 +27,6 @@ export interface PatternsBreakdownSceneState extends SceneObjectState {
   loading?: boolean;
   error?: string;
   blockingMessage?: string;
-  legendSyncPatterns: Set<string>;
 }
 
 export type PatternFrame = {
@@ -54,7 +45,6 @@ export class PatternsBreakdownScene extends SceneObjectBase<PatternsBreakdownSce
           variables: [new CustomVariable({ name: VAR_LABEL_GROUP_BY, defaultToAll: true, includeAll: true })],
         }),
       loading: true,
-      legendSyncPatterns: new Set(),
       ...state,
     });
 
@@ -134,106 +124,11 @@ export class PatternsBreakdownScene extends SceneObjectBase<PatternsBreakdownSce
 
     const patternFrames = this.buildPatterns(lokiPatterns);
 
-    const logExploration = sceneGraph.getAncestor(this, IndexScene);
-
     this.setState({
-      body: this.getSingleViewLayout(patternFrames, logExploration),
+      body: new PatternsFrameScene({
+        patternFrames,
+      }),
       loading: false,
-    });
-  }
-
-  private extendTimeSeriesLegendBus(vizPanel: VizPanel, context: PanelContext) {
-    const originalOnToggleSeriesVisibility = context.onToggleSeriesVisibility;
-
-    context.onToggleSeriesVisibility = (label: string, mode: SeriesVisibilityChangeMode) => {
-      originalOnToggleSeriesVisibility?.(label, mode);
-
-      const override: ConfigOverrideRule | undefined = vizPanel.state.fieldConfig.overrides?.[0];
-      const patternsToShow: string[] = override?.matcher.options.names;
-      const legendSyncPatterns = new Set<string>();
-
-      if (patternsToShow) {
-        patternsToShow.forEach(legendSyncPatterns.add, legendSyncPatterns);
-      }
-
-      this.setState({
-        legendSyncPatterns,
-      });
-    };
-  }
-
-  private getSingleViewLayout(patternFrames: PatternFrame[], logExploration: IndexScene) {
-    const appliedPatterns = sceneGraph.getAncestor(logExploration, IndexScene).state.patterns;
-    const timeRange = sceneGraph.getTimeRange(this).state.value;
-
-    const timeSeries = PanelBuilders.timeseries()
-      .setData(
-        new SceneDataNode({
-          data: {
-            series: patternFrames.map((patternFrame, seriesIndex) => {
-              // Mutating the dataframe config here means that we don't need to update the colors in the table view
-              const dataFrame = patternFrame.dataFrame;
-              dataFrame.fields[1].config.color = overrideToFixedColor(seriesIndex);
-              return dataFrame;
-            }),
-            state: LoadingState.Done,
-            timeRange: timeRange,
-          },
-        })
-      )
-      .setOption('legend', {
-        asTable: true,
-        showLegend: true,
-        displayMode: LegendDisplayMode.Table,
-        placement: 'right',
-        width: 200,
-      })
-      .setLinks([
-        {
-          url: '#',
-          targetBlank: false,
-          onClick: (event) => {
-            onPatternClick({
-              pattern: event.origin.name,
-              type: 'include',
-              indexScene: logExploration,
-            });
-          },
-          title: 'Include',
-        },
-        {
-          url: '#',
-          targetBlank: false,
-          onClick: (event) => {
-            onPatternClick({
-              pattern: event.origin.name,
-              type: 'exclude',
-              indexScene: logExploration,
-            });
-          },
-          title: 'Exclude',
-        },
-      ])
-      .build();
-
-    timeSeries.setState({
-      extendPanelContext: (vizPanel, context) => this.extendTimeSeriesLegendBus(vizPanel, context),
-    });
-
-    return new SceneCSSGridLayout({
-      templateColumns: '100%',
-
-      children: [
-        new SceneFlexItem({
-          minHeight: 200,
-          maxWidth: '100%',
-          body: timeSeries,
-        }),
-        new PatternsViewTableScene({
-          patternFrames,
-          appliedPatterns,
-        }),
-      ],
     });
   }
 
@@ -344,11 +239,4 @@ export function buildPatternsScene() {
   return new SceneFlexItem({
     body: new PatternsBreakdownScene({}),
   });
-}
-
-export function overrideToFixedColor(key: keyof typeof palette): FieldColor {
-  return {
-    mode: 'fixed',
-    fixedColor: palette[key] as string,
-  };
 }
